separated hd_node from mnemonic.ts
