+++ packages/jellyfish-wallet-mnemonic/src/index.ts
@@ -1,1 +1,2 @@
+export * from './hd_node'
 export * from './mnemonic'
